extract HighlightedText component into its own file
